🐛 修复 Toaster 不跟随用户选择的明暗主题
@opskat/ui 的 Toaster 从 next-themes 的 useTheme 取主题，但应用挂的是自己的
ThemeProvider，从没挂过 NextThemesProvider——useTheme 在 Provider 外返回
undefined，解构默认值把它兜成 "system"，sonner 于是按 prefers-color-scheme
自行解析。结果：用户在深色系统下显式选浅色（或相反）时，toast 跟的是系统而
不是用户的选择。

主题的真相是 <html> 上的 .dark class（ThemeProvider 写入，已解析过 system
并监听系统变化）。应用侧本就有 useResolvedTheme 观察它，但它在 src/ 下，
包不能反向依赖，照抄一份又会立刻分叉——所以把实现移进 @opskat/ui，
theme-provider.tsx 改为重新导出，五处既有 import 路径不变。

next-themes 是 Next.js 的库，移除后包内已无引用，一并从依赖与 lockfile 删除。

测试先行：ToasterTheme.test.tsx 断言 .dark 下渲染 dark、无 .dark 下渲染
light、显式 theme 优先。修复前第一条失败（实际渲染 light）。

docs/DESIGN.md 同步：richColors 实际由 App.tsx 挂载处传入，不在 sonner.tsx。
