+++ frontend/packages/ui/package.json
@@ -20,7 +20,6 @@
     "sonner": "^2.0.7",
     "tailwind-merge": "^3.0.0",
     "lucide-react": "^0.400.0",
-    "next-themes": "^0.4.6",
     "pinyin-pro": "^3.28.0"
   }
 }
